requirements.txt
speedtest-cli
pythonping
flask
